requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-groq==0.0.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.1
nltk==3.8.1 